fix, added favorite view model
