+++ app/build.gradle
@@ -65,7 +65,7 @@ dependencies {
     implementation libraries.roomRuntime
     implementation libraries.roomKtx
     kapt libraries.roomCompiler
-    implementation 'androidx.lifecycle:lifecycle-livedata-ktx:2.5.1'
+    implementation libraries.livedata
 
 //    androidTestImplementation"androidx.room:room-testing:$rootProject.roomVersion"
 //    implementation "androidx.room:room-rxjava2:$rootProject.roomVersion"
