+++ app/src/main/java/ru/android/myrecipesbook/ui/fragment/FavoriteFragment.kt
@@ -6,8 +6,14 @@ import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.fragment.app.Fragment
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.Observer
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
 import androidx.recyclerview.widget.LinearLayoutManager
+import androidx.recyclerview.widget.RecyclerView
 import retrofit2.Call
 import retrofit2.Response
 import ru.android.myrecipesbook.R
@@ -16,11 +22,13 @@ import ru.android.myrecipesbook.adapter.DishAdapterVertical
 import ru.android.myrecipesbook.databinding.FragmentFavoriteBinding
 import ru.android.myrecipesbook.model.RecipesResponse
 import ru.android.myrecipesbook.repository.FakeFoodRepository
+import ru.android.myrecipesbook.ui.viewmodel.FavoriteViewModel
 import timber.log.Timber
 
 
 class FavoriteFragment : Fragment() {
 
+    private lateinit var favoriteViewModel: FavoriteViewModel
     private lateinit var binding: FragmentFavoriteBinding
     private var fakeFoodRepository = FakeFoodRepository
 
@@ -28,6 +36,8 @@ class FavoriteFragment : Fragment() {
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
+
+        favoriteViewModel = ViewModelProvider(this)[FavoriteViewModel::class.java]
         binding = FragmentFavoriteBinding.inflate(inflater, container, false)
         val root = binding.root
 
@@ -35,32 +45,16 @@ class FavoriteFragment : Fragment() {
 
         val recycleViewDish = binding.recycleDishVertical
         recycleViewDish.layoutManager = LinearLayoutManager(activity)
-//        val dish = fakeFoodRepository.getListOfDishes()
-//        recycleViewDish.adapter = DishAdapter(dish, R.layout.list_item_vertical_dish)
 
-        val getRecipes = RecipeApiClient.apiClient.getAllRecipes()
-
-        getRecipes.enqueue(object : retrofit2.Callback<List<RecipesResponse>> {
-
-            override fun onResponse(
-                call: Call<List<RecipesResponse>>,
-                response: Response<List<RecipesResponse>>
-            ) {
-                val dish = response.body()?.get(0)
-                recycleViewDish.adapter = dish?.let {
-                    DishAdapterVertical(
-                        it,
-                        R.layout.list_item_vertical_dish
-                    )
-                }
-                Log.d("", response.toString())
-            }
-
-            override fun onFailure(call: Call<List<RecipesResponse>>, t: Throwable) {
-                Log.e(TAG, t.toString())
-            }
-
+        favoriteViewModel.favoriteLiveData.observe(this.viewLifecycleOwner, Observer {
+            DishAdapterVertical(it, R.layout.list_item_vertical_dish)
+        }
+        )
+        favoriteViewModel.errorLiveData.observe(this.viewLifecycleOwner, Observer<String> { error ->
+            Toast.makeText(context, error, Toast.LENGTH_SHORT).show()
         })
+
+        favoriteViewModel.getRecipeFromNetwork(recycleViewDish)
 
         return root
     }
